🛠️ refactor: (FE) - 페이지 폴더구조 변경
